requirements.txt
discord.py[voice]
python-dotenv
openai
requests
google-genai
pillow
icecream
beautifulsoup4
boto3
tenacity
pytest